added initial version of slave sketch

+++ dev/platformio/slave_drum/src/RfMessage.h
@@ -3,11 +3,11 @@
  * @Date:   2018-07-22T20:22:48+02:00
  * @Project: Brain String
  * @Last modified by:   lutz
- * @Last modified time: 2018-07-22T20:23:22+02:00
+ * @Last modified time: 2018-07-22T20:52:59+02:00
  */
 
  struct RfMessage {
  	char type;
- 	uint8_t value;
+ 	uint16_t value;
  	uint8_t id;
  };
